chore : move the test file(StringCalculatorTest, StringValidatorTest) to domain package
